fix: skill tags
