+++ src/functions/constants/skills.ts
@@ -4,45 +4,5 @@ import {
   TOOL_SKILLS,
 } from './skillsData';
 
-export const SKILL_TAGS = {
-  TECHNICAL: {
-    JQUERY: 'JQuery',
-    REACT: 'React',
-    NEXT_JS: 'Next.js',
-    VUE_JS: 'Vue.js',
-    TYPESCRIPT: 'TypeScript',
-    TAILWIND: 'Tailwind',
-    HTML_PUG_EJS: 'HTML/Pug/Ejs',
-    CSS_SASS: 'CSS/Sass',
-    JAVASCRIPT: 'JavaScript',
-    PYTHON: 'Python',
-    LARAVEL: 'Laravel',
-    MOVABLE_TYPE: 'MovableType',
-    WORDPRESS: 'WordPress',
-    VITE: 'Vite',
-    WEBPACK: 'Webpack',
-    GULP: 'Gulp',
-  },
-  OTHERS_LIBRARY: {
-    RADIX_UI: 'Radix UI',
-    SHADCN_UI: 'shadcn/ui',
-    MATERIAL_UI: 'Material UI',
-    REACT_HOOK_FORM: 'react-hook-form',
-    ZOD: 'zod',
-    ESLINT: 'ESLint',
-    PRETTIER: 'Prettier',
-    STYLELINT: 'Stylelint',
-    MARKUPLINT: 'Markuplint',
-  },
-  TOOL: {
-    PHOTOSHOP: 'PhotoShop',
-    ILLUSTRATOR: 'Illustrator',
-    XD: 'XD',
-    FIGMA: 'Figma',
-    GITHUB: 'Github',
-    VSCODE: 'VSCode',
-    BACKLOG: 'Backlog',
-  },
-};
-
+export { SKILL_TAGS } from './skillTags';
 export { TECHNICAL_SKILLS, OTHERS_LIBRARY_SKILLS, TOOL_SKILLS };
