--- src/lib/data/articles.json
@@ -1,16 +1,0 @@
-[
-  {
-    "id": "1",
-    "title": "最初の記事",
-    "content": "これは最初の記事の内容です。",
-    "thumbnail": "profile.png",
-    "imagePath": ["profile.png", "profile.png"]
-  },
-  {
-    "id": "2",
-    "title": "2番目の記事",
-    "content": "これは2番目の記事の内容です。",
-    "thumbnail": "article2.png",
-    "imagePath": ["article2.jpg"]
-  }
-]
